requirements.txt
Django>=3.2,<5
djangorestframework>=3.12
django-cors-headers>=3.13
mysqlclient>=2.0